requirements/base.txt
amqp==5.3.1
asgiref==3.9.1
attrs==25.3.0
billiard==4.2.1
celery==5.5.3
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
dj-database-url==3.0.1
dj-rest-auth==7.0.1
django==5.2.5
django-anymail==13.1
django-cors-headers==4.7.0
django-filter==25.1
django-picklefield==3.3
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
dnspython==2.7.0
drf-spectacular==0.28.0
drf-writable-nested==0.7.2
email-validator==2.3.0
idna==3.10
inflection==0.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kombu==5.5.4
numpy==2.3.2
oauthlib==3.3.1
packaging==25.0
pandas==2.3.2
pillow==11.3.0
psycopg2-binary==2.9.11
pyjwt==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pyyaml==6.0.2
redis==3.5.3
referencing==0.36.2
requests==2.32.5
requests-oauthlib==2.0.0
rest-framework-simplejwt==0.0.2
rpds-py==0.27.0
six==1.17.0
sqlparse==0.5.3
typing-extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.6.1
vine==5.1.0
whitenoise==6.11.0
django-tasks==0.12.0
